created committee component and moved contentful call fxn along with content type
