+++ src/App.js
@@ -1,20 +1,21 @@
 import './App.css';
 import { useEffect, useState } from 'react';
 import useContentful from './useContentful';
-
+import Committee from './components/Committee';
 
 function App() {
 
-  const [members, setMembers] = useState([]);
-  const { getMembers } = useContentful();
+  // const [members, setMembers] = useState([]);
+  // const { getMembers } = useContentful();
 
-  useEffect(() => {
-    getMembers().then((res) => console.log(res));
-  })
+  // useEffect(() => {
+  //   getMembers({content_type: "steeringCommittee"}).then((res) => console.log(res.items));
+  // })
 
   return (
     <div className="App">
       <h1>working?</h1>
+      <Committee />
     </div>
   );
 }
